extract run id parsing
